created distance conversion module and exposed functions, not constants
