+++ src-tauri/src/aux_functions/mod.rs
@@ -1,5 +1,4 @@
 pub(crate) mod adj_matrix;
-pub(crate) mod conversion_factors;
 pub(crate) mod convert_to_graph_from_string;
 pub(crate) mod edge_factory;
 pub(crate) mod get_most_similar;
